app/requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==3.2.0
click==8.1.6
flake8==6.0.0
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.0
idna==3.4
magic-filter==1.0.10
mccabe==0.7.0
multidict==6.0.4
psycopg2==2.9.6
pycodestyle==2.10.0
pydantic==1.10.12
pydantic_core==2.3.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
redis==5.0.0
aiogram-dialog==2.0.0